+++ package.go
@@ -1,0 +1,4 @@
+/* Package qual implements tests and helpers for writing highly readable tests.
+
+ */
+package qual